pyproject.toml
[project]
name = "aiecommerce-agents"
version = "0.1.0"
description = "AI-powered e-commerce agent orchestration platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "langgraph>=0.4.1",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.24",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "aiosqlite>=0.21.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.4",
]

# ── Build system (src layout) ──────────────────────────────────────────
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

# ── Ruff (linter + formatter) ──────────────────────────────────────────
[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["orchestrator"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

# ── Mypy (static type checker) ─────────────────────────────────────────
[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["langgraph.*", "langchain.*", "langchain_openai.*"]
ignore_missing_imports = true

# ── Pytest ─────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/orchestrator --cov-report=term-missing"
pythonpath = ["src"]
